Joined into one class buildIn and transfer type definition parser expanders
